+++ packages/gatsby/src/utils/webpack.config.js
@@ -361,6 +361,9 @@ module.exports = async (
         // Using directories for module resolution is mandatory because
         // relative path imports are used sometimes
         // See https://stackoverflow.com/a/49455609/6420957 for more details
+        "@babel/runtime": path.dirname(
+          require.resolve(`@babel/runtime/package.json`)
+        ),
         "core-js": path.dirname(require.resolve(`core-js/package.json`)),
         "react-hot-loader": path.dirname(
           require.resolve(`react-hot-loader/package.json`)
@@ -432,6 +435,52 @@ module.exports = async (
       },
       minimize: !program.noUglify,
     }
+  }
+
+  if (stage === `build-html` || stage === `develop-html`) {
+    const externalList = [
+      // match `lodash` and `lodash/foo`
+      // but not things like `lodash-es`
+      `lodash`,
+      /^lodash\//,
+      `react`,
+      /^react-dom\//,
+      `pify`,
+      `@reach/router`,
+      `@reach/router/lib/history`,
+      `common-tags`,
+      `path`,
+      `semver`,
+      `react-helmet`,
+      `minimatch`,
+      `fs`,
+      /^core-js\//,
+      `es6-promise`,
+      `crypto`,
+      `zlib`,
+      `http`,
+      `https`,
+      `debug`,
+    ]
+
+    config.externals = [
+      function(context, request, callback) {
+        if (
+          externalList.some(item => {
+            if (typeof item === `string` && item === request) {
+              return true
+            } else if (item instanceof RegExp && item.test(request)) {
+              return true
+            }
+
+            return false
+          })
+        ) {
+          return callback(null, `umd ${request}`)
+        }
+        return callback()
+      },
+    ]
   }
 
   store.dispatch(actions.replaceWebpackConfig(config))